feat: adds browser url construction from uri
